+++ counting-sort/countingSort.h
@@ -5,6 +5,6 @@
 #include "time.h"
 using namespace std;
 void countingSort(int a[], int n);
-void countingSortCompare(int a[], int n, int &comparison);
+void countingSortCompare(int a[], int n, long long &comparison);
 void countingSortTime(int a[], int n, double &time);
 #endif
